fix(api): preserve original TextToAudioPayload validation per endpoint
The shared TextToAudioPayload added UUID validation on message_id,
but console/explore originally had no such validation. Move the UUID
validator to a web-specific subclass to avoid the behavioral regression.
